demo/requirements.txt
pathway==0.8.4
streamlit==1.31.0
llama-index==0.9.38
python-dotenv==1.0.1
traceloop-sdk==0.12.5
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp-proto-grpc>=1.22.0
